requirements.txt
django>=5.0,<6.1
djangorestframework>=3.14
psycopg2-binary>=2.9
pillow>=10.0
python-decouple>=3.8
django-cors-headers>=4.3
